More template-based LD opcodes.
add: LD_0_3__1
add: *GetReg*-based template test to have a single function aliasing a register, to be used by both the read and write Impls.  The unspecialized template method can call the GetReg method (which will not have a specialization for any indirect addressing), and an Impl specialization could be made to handle said indirect addressing.