Activate rough memory bus caching scheme, boosts debug performance tremendously (back up to release speeds).
change: try to cache memory bus devices in the case where the first few big if's in the memory bus fall through.  If the cache fails, just defer to the old code.  Eventually this will be cleaned up and the caching will be required to pass, etc. but in the meantime the speed boost is much appreciated.