Sound: add CH1, without frequency sweep.
fix: channel turns off when a length of 0 is written to NRx1 (docs are unclear on this point but it seems to make sense)
fix: device is not unpaused before Sound is reset (I meant to remove this earlier but somehow forgot - the code to unpause the device after reset is already there).  Fixes rare crash on startup.